+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - `Innmind\Xml\Attribute::toString()`
 - `Innmind\Xml\Visitor\Text`
 - `Innmind\Xml\Document::toString()`, use `->asContent(Format::inline)->toString()` instead
+- `Innmind\Xml\Document::content()`
 
 ### Fixed